+++ NEWS.md
@@ -1,6 +1,12 @@
+# scorecard 0.3.3
+
+* the perf_psi function modified the results from psi to csi for variables.
+* add the global argument `options(scorecard.bin_close_right  = TRUE)` to set the bins closed on the right.
+* fixed a bug in describe function.
+
 # scorecard 0.3.2
 
-* added a functon `describe` that calculatess statistic parameters for exploratory data analysis.
+* added a function `describe` that calculatess statistic parameters for exploratory data analysis.
 * replace good/bad with neg/pos.
 * replace NaN with NA if it exists in the input data.
 * addd [] after the returned DataTable in gains_table and woebin function in order to directly display the datatable when printing.